fix: remove options working

+++ wp-tarteaucitron/uninstall.php
@@ -23,7 +23,9 @@ wp_tarteaucitron_uninstall();
 function wp_tarteaucitron_uninstall(): void {
 	try {
 		wordpress_plugin_uninstall_called();
-		delete_options();
+		if(get_option('wp_tarteaucitron_remove_options')){
+			delete_options();
+		}
 	} catch( Exception $exception ) {
 		exit( $exception->getMessage() );
 	}
@@ -56,6 +58,11 @@ function wordpress_plugin_uninstall_called(): bool {
 function delete_options(): void {
 	delete_option( 'wp_tarteaucitron_privacy_policy_url' );
 	delete_option( 'wp_tarteaucitron_use_wp_privacy_policy_page' );
+	delete_option( 'wp_tarteaucitron_remove_credit' );
+	delete_option( 'wp_tarteaucitron_icon_position' );
+	delete_option( 'wp_tarteaucitron_privacy_cookie_name' );
+	delete_option( 'wp_tarteaucitron_hashtag' );
+	delete_option( 'wp_tarteaucitron_remove_options' );
 }
 
 ?>
